[sim] Bug fixes in shifts, plus a new test case

+++ riscv/insns/srlvw.h
@@ -1,1 +1,1 @@
-RC = sext32(RB >> (RA & 0x1F));
+RC = sext32((uint32_t)RB >> (RA & 0x1F));
